refactor[FE]: Move player matchtable to separete file
